feat(workflow): migrate http, code and list to the catalog
