+++ apps/web/src/components/workflow/nodes/core/code/schema.ts
@@ -1,158 +1,42 @@
 // apps/web/src/components/workflow/nodes/core/code/schema.ts
 
-import { z } from 'zod'
 import {
-  BaseType,
-  NodeCategory,
-  type NodeDefinition,
-  NodeType,
-  type UnifiedVariable,
-  type ValidationResult,
-} from '~/components/workflow/types'
+  codeManifest,
+  type NodeManifest,
+  normalizeCodeOutputType,
+} from '@auxx/lib/workflow-engine/client'
+import { BaseType, type NodeDefinition, type UnifiedVariable } from '~/components/workflow/types'
 import { createUnifiedOutputVariable } from '~/components/workflow/utils/variable-conversion'
+import { defineFromManifest } from '../../define-from-manifest'
 import type { CodeNodeData } from './types'
 
-/**
- * Zod schema for code configuration
- * @deprecated Use codeNodeDataSchema instead
- */
-export const codeSchema = z.object({
-  title: z.string().min(1),
-  desc: z.string().optional(),
-  // variables: z
-  //   .array(z.object({ variable: z.string(), value_selector: z.array(z.string()) }))
-  //   .default([]),
-  code_language: z.enum(['javascript', 'python3']).default('javascript'),
-  code: z.string(),
-  inputs: z.array(z.object({ name: z.string(), variableId: z.string() })).optional(),
-  outputs: z
-    .array(z.object({ name: z.string(), type: z.any(), description: z.string().optional() }))
-    .optional(),
-  // legacyOutputs: z.record(z.string(), z.any()).optional(),
-})
+// The data half (data interface, zod schema, defaults, validator, variable
+// extraction) lives in the node catalog
+// (`@auxx/lib/workflow-engine/catalog/nodes/code`). This file is the merge
+// site: manifest + the web-only output resolver.
 
 /**
- * Zod schema for code node data (flattened structure)
- */
-export const codeNodeDataSchema = z.object({
-  // Base node properties
-  id: z.string(),
-  type: z.literal(NodeType.CODE),
-  selected: z.boolean(),
-
-  // Flattened config properties
-  title: z.string().min(1),
-  desc: z.string().optional(),
-  code_language: z.enum(['javascript', 'python3']).default('javascript'),
-  code: z.string(),
-  inputs: z.array(z.object({ name: z.string(), variableId: z.string() })).optional(),
-  outputs: z
-    .array(z.object({ name: z.string(), type: z.any(), description: z.string().optional() }))
-    .optional(),
-  // legacyOutputs: z.record(z.string(), z.any()).optional(),
-
-  // Other node data properties
-  isValid: z.boolean().optional(),
-  errors: z.array(z.string()).optional(),
-  disabled: z.boolean().optional(),
-  outputVariables: z.array(z.string()).optional(),
-})
-
-/**
- * Default data for new code nodes (flattened)
- */
-export const codeDefaultData: Partial<CodeNodeData> = {
-  title: 'Code',
-  desc: 'Transform data with code',
-  variables: [],
-  code_language: 'javascript',
-  code: `const main = async () => {
-  // You can use input variables here if defined
-  return {
-    output1: undefined
-  }
-}`,
-  inputs: [],
-  outputs: [
-    { name: 'output1', type: BaseType.STRING, description: 'First output from the code execution' },
-  ],
-}
-
-/**
- * Validation function for code configuration
- */
-export const validateCodeConfig = (data: CodeNodeData): ValidationResult => {
-  const errors: Array<{ field: string; message: string; type?: 'warning' | 'error' }> = []
-
-  // Validate title
-  if (!data.title?.trim()) {
-    errors.push({ field: 'title', message: 'Title is required', type: 'error' })
-  }
-
-  // Validate code content
-  if (!data.code?.trim()) {
-    errors.push({ field: 'code', message: 'Code is required', type: 'error' })
-  }
-
-  // Basic syntax check for JavaScript
-  if (data.code_language === 'javascript') {
-    try {
-      new Function(data.code)
-
-      // Check for main function definition
-      if (data.code && !data.code.includes('main')) {
-        errors.push({
-          field: 'code',
-          message: 'Code must define a main() function',
-          type: 'error',
-        })
-      }
-
-      // Add warnings for potentially unsafe patterns
-      if (data.code?.includes('eval(')) {
-        errors.push({
-          field: 'code',
-          message: 'Using eval() is potentially unsafe and should be avoided',
-          type: 'warning',
-        })
-      }
-
-      if (data.code?.includes('innerHTML')) {
-        errors.push({
-          field: 'code',
-          message: 'Direct innerHTML manipulation can lead to XSS vulnerabilities',
-          type: 'warning',
-        })
-      }
-    } catch (error) {
-      errors.push({
-        field: 'code',
-        message: `JavaScript syntax error: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        type: 'error',
-      })
-    }
-  }
-
-  const hasErrors = errors.filter((e) => e.type === 'error').length > 0
-  return { isValid: !hasErrors, errors }
-}
-
-/**
- * Define output variables for code node
+ * Define output variables for code node.
+ *
+ * Drift fix (node-catalog Phase 1, plan §6): this used to read
+ * `output.type?.type` — an object shape only the legacy `CodeOutput` map ever
+ * used — while the output editor writes a plain `BaseType` string, so every
+ * output resolved as STRING regardless of what the user picked.
+ * `normalizeCodeOutputType` reads the string first and tolerates the legacy
+ * object shape. User-visible: code-node outputs now show their configured type
+ * in the variable picker.
  */
 const getCodeOutputVariables = (data: CodeNodeData, nodeId: string): UnifiedVariable[] => {
   const outputs: UnifiedVariable[] = []
 
   // Use new outputs format if available
   if (data.outputs && Array.isArray(data.outputs)) {
-    data.outputs.forEach((output: any) => {
-      const baseType = (output.type?.type as BaseType) || BaseType.STRING
-
+    data.outputs.forEach((output) => {
       outputs.push(
         createUnifiedOutputVariable({
           nodeId,
-          path: output.name, // Changed from 'name' to 'path'
-          type: baseType,
+          path: output.name,
+          type: normalizeCodeOutputType(output.type),
           description: output.description || `Output: ${output.name}`,
         })
       )
@@ -162,7 +46,7 @@ const getCodeOutputVariables = (data: CodeNodeData, nodeId: string): UnifiedVari
     outputs.push(
       createUnifiedOutputVariable({
         nodeId,
-        path: 'output1', // Changed from 'name' to 'path'
+        path: 'output1',
         type: BaseType.OBJECT,
         description: 'Result from code execution',
       })
@@ -172,32 +56,18 @@ const getCodeOutputVariables = (data: CodeNodeData, nodeId: string): UnifiedVari
   return outputs
 }
 
-export function extractCodeVariables(data: CodeNodeData): string[] {
-  const uniqueVariables = new Set<string>()
+/** Node definition for code */
+export const codeDefinition: NodeDefinition<CodeNodeData> = defineFromManifest(
+  codeManifest as unknown as NodeManifest<CodeNodeData>,
+  { outputVariables: getCodeOutputVariables }
+)
 
-  data.inputs?.forEach((input: any) => {
-    if (input.variableId) {
-      uniqueVariables.add(input.variableId)
-    }
-  })
+// Back-compat re-exports so no consumer import churns:
+export {
+  codeNodeDataSchema,
+  extractCodeVariables,
+  validateCodeConfig,
+} from '@auxx/lib/workflow-engine/client'
 
-  return Array.from(uniqueVariables)
-}
-
-/**
- * Node definition for code
- */
-export const codeDefinition: NodeDefinition<CodeNodeData> = {
-  id: NodeType.CODE,
-  category: NodeCategory.TRANSFORM,
-  displayName: 'Code',
-  description: 'Execute custom code to transform data',
-  icon: 'code',
-  color: '#8B5CF6', // TRANSFORM category color
-  defaultData: codeDefaultData,
-  schema: codeNodeDataSchema,
-  validator: validateCodeConfig as any,
-  canRunSingle: true,
-  extractVariables: extractCodeVariables as any,
-  outputVariables: getCodeOutputVariables as any,
-}
+/** Default data for new code nodes (flattened) */
+export const codeDefaultData = codeManifest.defaultData() as Partial<CodeNodeData>
