+++ packages/lib/src/signals/email/__tests__/tracking-tokens.test.ts
@@ -64,7 +64,7 @@ describe('tracking-tokens', () => {
     // carries only 2 significant bits of a 32-byte HS256 signature (43 base64url
     // chars encode 258 bits), so several values there decode to identical bytes
     // and "tampering" with it hands the verifier a still-valid token.
-    const [header, payload, signature] = token.split('.')
+    const [header, payload, signature = ''] = token.split('.')
     const flipped = `${signature.startsWith('A') ? 'B' : 'A'}${signature.slice(1)}`
     const tampered = `${header}.${payload}.${flipped}`
